requirements.txt
# Django libraries
django==3.0.5
django-rest-framework==0.1.0
django-extensions==2.2.9